+++ src/combi.h
@@ -10,6 +10,6 @@ namespace GMDH {
 
         GmdhModel& fit(const MatrixXd& x, const VectorXd& y,
                        const Criterion& criterion = Criterion(CriterionType::regularity),
-                       double testSize = 0.5, uint8_t pAverage = 1, int threads = 1, int verbose = 0, double limit = 0);
+                       double testSize = 0.5, int pAverage = 1, int threads = 1, int verbose = 0, double limit = 0);
     };
 }
